requirements.txt
alabaster==0.7.12
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.9
docutils==0.14
en-core-web-sm==2.0.0
idna==2.8
imagesize==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nose==1.3.7
numpy==1.16.1
packaging==19.0
pdfminer.six==20181108
plac==0.9.6
preshed==2.0.1
pycryptodome==3.7.3
Pygments==2.3.1
pyparsing==2.3.1
pytz==2018.9
regex==2018.1.10
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
sortedcontainers==2.1.0
spacy==2.0.18
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
thinc==6.12.1
toolz==0.9.0
tqdm==4.31.1
ujson==1.35
urllib3==1.24.1
wrapt==1.10.11
